+++ src/client/api.rs
@@ -21,164 +21,7 @@ use crate::jce::{SvcDevLoginInfo, SvcRespRegister};
 use crate::{RQError, RQResult};
 
 /// 登录相关
-impl super::Client {
-    /// 二维码登录 - 获取二维码
-    pub async fn fetch_qrcode(&self) -> RQResult<QRCodeState> {
-        let req = self.engine.read().await.build_qrcode_fetch_request_packet();
-        let resp = self.send_and_wait(req).await?;
-        let resp = self
-            .engine
-            .read()
-            .await
-            .decode_trans_emp_response(resp.body)?;
-        self.process_trans_emp_response(resp.clone()).await;
-        Ok(resp)
-    }
-
-    /// 二维码登录 - 查询二维码状态
-    pub async fn query_qrcode_result(&self, sig: &[u8]) -> RQResult<QRCodeState> {
-        let req = self
-            .engine
-            .read()
-            .await
-            .build_qrcode_result_query_request_packet(sig);
-        let resp = self.send_and_wait(req).await?;
-        let resp = self
-            .engine
-            .read()
-            .await
-            .decode_trans_emp_response(resp.body)?;
-        self.process_trans_emp_response(resp.clone()).await;
-        Ok(resp)
-    }
-
-    /// 二维码登录 - 登录 ( 可能还需要 device_lock_login )
-    pub async fn qrcode_login(
-        self: &Arc<Self>,
-        tmp_pwd: &[u8],
-        tmp_no_pic_sig: &[u8],
-        tgt_qr: &[u8],
-    ) -> RQResult<LoginResponse> {
-        let req =
-            self.engine
-                .read()
-                .await
-                .build_qrcode_login_packet(tmp_pwd, tmp_no_pic_sig, tgt_qr);
-        let resp = self.send_and_wait(req).await?;
-        let resp = self.engine.read().await.decode_login_response(resp.body)?;
-        self.process_login_response(resp.clone()).await;
-        Ok(resp)
-    }
-
-    /// 密码登录 - 提交密码md5
-    pub async fn password_md5_login(
-        self: &Arc<Self>,
-        uin: i64,
-        password_md5: &[u8],
-    ) -> RQResult<LoginResponse> {
-        self.engine.read().await.uin.store(uin, Ordering::Relaxed);
-        let req = self
-            .engine
-            .read()
-            .await
-            .build_login_packet(password_md5, true);
-        let resp = self.send_and_wait(req).await?;
-        let resp = self.engine.read().await.decode_login_response(resp.body)?;
-        self.process_login_response(resp.clone()).await;
-        Ok(resp)
-    }
-
-    pub async fn password_login(
-        self: &Arc<Self>,
-        uin: i64,
-        password: &str,
-    ) -> RQResult<LoginResponse> {
-        self.password_md5_login(uin, &md5::compute(password).to_vec())
-            .await
-    }
-
-    /// 密码登录 - 请求短信验证码
-    pub async fn request_sms(self: &Arc<Self>) -> RQResult<LoginResponse> {
-        let req = self.engine.read().await.build_sms_request_packet();
-        let resp = self.send_and_wait(req).await?;
-        let resp = self.engine.read().await.decode_login_response(resp.body)?;
-        self.process_login_response(resp.clone()).await;
-        Ok(resp)
-    }
-
-    /// 密码登录 - 提交短信验证码
-    pub async fn submit_sms_code(self: &Arc<Self>, code: &str) -> RQResult<LoginResponse> {
-        let req = self
-            .engine
-            .read()
-            .await
-            .build_sms_code_submit_packet(code.trim());
-        let resp = self.send_and_wait(req).await?;
-        let resp = self.engine.read().await.decode_login_response(resp.body)?;
-        self.process_login_response(resp.clone()).await;
-        Ok(resp)
-    }
-
-    /// 密码登录 - 提交滑块ticket
-    pub async fn submit_ticket(self: &Arc<Self>, ticket: &str) -> RQResult<LoginResponse> {
-        let req = self.engine.read().await.build_ticket_submit_packet(ticket);
-        let resp = self.send_and_wait(req).await?;
-        let resp = self.engine.read().await.decode_login_response(resp.body)?;
-        self.process_login_response(resp.clone()).await;
-        Ok(resp)
-    }
-
-    /// 设备锁登录 - 二维码、密码登录都需要
-    pub async fn device_lock_login(self: &Arc<Self>) -> RQResult<LoginResponse> {
-        let req = self.engine.read().await.build_device_lock_login_packet();
-        let resp = self.send_and_wait(req).await?;
-        let resp = self.engine.read().await.decode_login_response(resp.body)?;
-        self.process_login_response(resp.clone()).await;
-        Ok(resp)
-    }
-
-    /// token 登录
-    pub async fn token_login(self: &Arc<Self>, mut token: impl Buf) -> RQResult<()> {
-        self.load_token(&mut token).await;
-        let req = self.engine.read().await.build_request_change_sig_packet();
-        self.send_and_wait(req).await?;
-        self.register_client().await?;
-        let r = tokio::join! {
-            self.wait_packet("StatSvc.ReqMSFOffline", 1),
-            self.wait_packet("MessageSvc.PushForceOffline", 1)
-        };
-        if let (Err(RQError::Timeout), Err(RQError::Timeout)) = r {
-            self.handler
-                .handle(QEvent::LoginEvent(self.uin().await))
-                .await;
-            Ok(())
-        } else {
-            Err(RQError::TokenLoginFailed)
-        }
-    }
-
-    /// 注册客户端，登录后必须注册
-    pub async fn register_client(&self) -> RQResult<SvcRespRegister> {
-        let req = self.engine.read().await.build_client_register_packet();
-        let resp = self.send_and_wait(req).await?;
-        let resp = self
-            .engine
-            .read()
-            .await
-            .decode_client_register_response(resp.body)?;
-        if !resp.result.is_empty() || resp.reply_code != 0 {
-            return Err(RQError::Other(resp.result + &resp.reply_code.to_string()));
-        }
-        self.online.store(true, Ordering::SeqCst);
-        Ok(resp)
-    }
-
-    pub async fn heartbeat(&self) -> RQResult<()> {
-        let req = self.engine.read().await.build_heartbeat_packet();
-        let _ = self.send_and_wait(req).await?;
-        Ok(())
-    }
-}
+mod logins;
 
 /// API
 impl super::Client {
